# openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Recipe API
  description: コーヒーレシピ管理システムのAPI仕様書
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: 開発環境

paths:
  /api/recipes:
    get:
      summary: レシピ検索
      description: 様々な条件でコーヒーレシピを検索・フィルタリングします
      operationId: searchRecipes
      tags:
        - Recipes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/RoastLevelParam'
        - $ref: '#/components/parameters/EquipmentParam'
        - $ref: '#/components/parameters/EquipmentTypeParam'
        - $ref: '#/components/parameters/TagsParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: レシピリストの取得に成功
          headers:
            X-Total-Count:
              description: 総件数
              schema:
                type: integer
            X-Request-ID:
              description: リクエスト追跡ID
              schema:
                type: string
                example: 'req_abc123xyz789'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeListResponse'
              examples:
                success:
                  summary: 成功例
                  value:
                    recipes:
                      - id: '1'
                        title: 'エチオピア イルガチェフェ V60'
                        summary: 'フルーティーで華やかな酸味が特徴的なエチオピア産コーヒー'
                        equipment: ['V60', 'コーヒーミル', 'ペーパーフィルター']
                        roastLevel: 'LIGHT_MEDIUM'
                        remarks: '特に蒸らし時間に注意。豆の膨らみを見て判断する。'
                        createdAt: '2025-06-20T10:30:00Z'
                        updatedAt: '2025-06-20T10:30:00Z'
                        tags:
                          - id: '1'
                            name: 'エチオピア'
                            slug: 'ethiopia'
                          - id: '2'
                            name: 'フルーティー'
                            slug: 'fruity'
                        baristaName: 'コーヒー愛好家'
                    pagination:
                      currentPage: 1
                      totalPages: 5
                      totalItems: 98
                      itemsPerPage: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: レシピ作成
      description: 新しいコーヒーレシピを作成します
      operationId: createRecipe
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeInput'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/recipes/{id}:
    get:
      summary: レシピ詳細取得
      description: 指定されたIDのレシピ詳細を取得します
      operationId: getRecipe
      tags:
        - Recipes
      parameters:
        - name: id
          in: path
          required: true
          description: レシピID（正の整数）
          schema:
            type: string
            pattern: '^[1-9][0-9]*$'
            example: '123'
      responses:
        '200':
          description: レシピ詳細の取得に成功
          headers:
            X-Request-ID:
              description: リクエスト追跡ID
              schema:
                type: string
                example: 'req_abc123xyz789'
            X-View-Count:
              description: 更新後のビュー数
              schema:
                type: integer
                example: 151
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: レシピ更新
      description: 指定されたIDのレシピを更新します
      operationId: updateRecipe
      tags:
        - Recipes
      parameters:
        - name: id
          in: path
          required: true
          description: レシピID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeInput'
      responses:
        '200':
          description: レシピの更新に成功
          headers:
            X-Request-ID:
              description: リクエスト追跡ID
              schema:
                type: string
                example: 'req_abc123xyz789'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: レシピ削除
      description: 指定されたIDのレシピを削除します
      operationId: deleteRecipe
      tags:
        - Recipes
      parameters:
        - name: id
          in: path
          required: true
          description: レシピID
          schema:
            type: string
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/equipment:
    get:
      summary: 器具一覧取得
      description: コーヒー器具をタイプ別に取得します
      operationId: getEquipment
      tags:
        - Equipment
      responses:
        '200':
          description: 器具一覧の取得に成功
          headers:
            X-Request-ID:
              description: リクエスト追跡ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
              examples:
                success:
                  summary: 成功例
                  value:
                    equipment:
                      grinder:
                        - id: '1'
                          name: 'バラッツァ エンコア'
                          brand: 'Baratza'
                          type: 'コーヒーミル'
                        - id: '2'
                          name: 'ハリオ コーヒーミル・セラミックスリム'
                          brand: 'HARIO'
                          type: 'コーヒーミル'
                      dripper:
                        - id: '3'
                          name: 'V60'
                          brand: 'HARIO'
                          type: 'ドリッパー'
                        - id: '4'
                          name: 'カリタ ウェーブドリッパー'
                          brand: 'Kalita'
                          type: 'ドリッパー'
                      filter:
                        - id: '5'
                          name: 'V60ペーパーフィルター'
                          brand: 'HARIO'
                          type: 'ペーパーフィルター'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    PageParam:
      name: page
      in: query
      description: ページ番号（1から開始）
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1

    LimitParam:
      name: limit
      in: query
      description: 1ページあたりのアイテム数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20

    SearchParam:
      name: search
      in: query
      description: レシピのタイトルや概要での部分一致検索キーワード
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 100

    RoastLevelParam:
      name: roastLevel
      in: query
      description: 焙煎度のフィルタ（カンマ区切りで複数指定可能）
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/RoastLevel'

    EquipmentParam:
      name: equipment
      in: query
      description: 器具名のフィルタ（カンマ区切りで複数指定可能）
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
          minLength: 1
          maxLength: 50

    EquipmentTypeParam:
      name: equipmentType
      in: query
      description: 器具タイプのフィルタ（カンマ区切りで複数指定可能）
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
          enum: [grinder, dripper, filter]

    TagsParam:
      name: tags
      in: query
      description: |
        タグIDによるフィルタ（カンマ区切りで複数指定可能）
        - 複数指定時はAND条件（すべてのタグに一致するレシピを返す）
        - 存在しないIDは無視される
        - 空の場合はフィルタなし
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
          pattern: '^[0-9]+$'

    SortParam:
      name: sort
      in: query
      description: ソートフィールド
      required: false
      schema:
        type: string
        enum: [id, title, viewCount, createdAt, updatedAt, publishedAt]
        default: id

    OrderParam:
      name: order
      in: query
      description: ソート順序
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc

  responses:
    Created:
      description: リソースの作成に成功
      headers:
        Location:
          description: 作成されたリソースのURL
          schema:
            type: string
            example: '/api/recipes/123'
        X-Request-ID:
          description: リクエスト追跡ID
          schema:
            type: string
            example: 'req_abc123xyz789'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Recipe'

    NoContent:
      description: 正常に処理されました（レスポンスボディなし）
      headers:
        X-Request-ID:
          description: リクエスト追跡ID
          schema:
            type: string
            example: 'req_abc123xyz789'

    BadRequest:
      description: リクエストパラメータが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_parameter:
              summary: 無効なパラメータ
              value:
                error: 'INVALID_PARAMETER'
                message: 'パラメータが不正です'
                request_id: 'req_abc123xyz789'
                timestamp: '2025-06-22T10:30:00Z'
            validation_error:
              summary: バリデーションエラー
              value:
                error: 'VALIDATION_ERROR'
                message: '入力データにエラーがあります'
                request_id: 'req_abc123xyz789'
                timestamp: '2025-06-22T10:30:00Z'
                details:
                  fields:
                    - field: 'title'
                      message: 'タイトルは必須です'
                    - field: 'roastLevel'
                      message: '焙煎度は必須です'
                    - field: 'beanWeight'
                      message: '豆の重量は1以上100以下で入力してください'
                    - field: 'waterTemp'
                      message: '湯温は60以上100以下で入力してください'
            missing_field:
              summary: 必須フィールド不足
              value:
                error: 'MISSING_REQUIRED_FIELD'
                message: '必須フィールドが不足しています'
                request_id: 'req_abc123xyz789'
                timestamp: '2025-06-22T10:30:00Z'
                details:
                  fields:
                    - field: 'title'
                      message: 'タイトルは必須です'

    Forbidden:
      description: 非公開レシピへのアクセス拒否
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unpublished_recipe:
              summary: 非公開レシピ
              value:
                error: 'RECIPE_NOT_PUBLISHED'
                message: 'このレシピは非公開です'
                request_id: 'req_abc123xyz789'
                timestamp: '2025-06-22T10:30:00Z'

    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resource_not_found:
              summary: リソース未発見
              value:
                error: 'RESOURCE_NOT_FOUND'
                message: '指定されたリソースが見つかりません'
                request_id: 'req_abc123xyz789'
                timestamp: '2025-06-22T10:30:00Z'

    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'INTERNAL_SERVER_ERROR'
            message: 'サーバー内部でエラーが発生しました'
            request_id: 'req_abc123xyz789'
            timestamp: '2025-06-22T10:30:00Z'

  schemas:
    Recipe:
      type: object
      description: コーヒーレシピ情報
      required:
        - id
        - title
        - roastLevel
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: レシピの一意識別子
          example: '1'
        title:
          type: string
          description: レシピのタイトル
          minLength: 1
          maxLength: 100
          example: 'エチオピア イルガチェフェ V60'
        summary:
          type: string
          description: レシピの概要説明
          minLength: 1
          maxLength: 500
          example: 'フルーティーで華やかな酸味が特徴的なエチオピア産コーヒー'
          nullable: true
        equipment:
          type: array
          description: 使用する器具のリスト
          items:
            type: string
            minLength: 1
            maxLength: 50
          minItems: 1
          maxItems: 10
          example: ['V60', 'コーヒーミル', 'ペーパーフィルター']
          nullable: true
        roastLevel:
          $ref: '#/components/schemas/RoastLevel'
        remarks:
          type: string
          description: 備考・コメント
          maxLength: 1000
          example: '特に蒸らし時間に注意。豆の膨らみを見て判断する。'
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 作成日時（ISO 8601形式）
          example: '2025-06-20T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時（ISO 8601形式）
          example: '2025-06-22T10:30:00Z'
        tags:
          type: array
          description: レシピに関連するタグ
          items:
            $ref: '#/components/schemas/RecipeTag'
          example:
            - id: '1'
              name: 'エチオピア'
              slug: 'ethiopia'
        baristaName:
          type: string
          description: バリスタ名（投稿者名）
          example: 'テスト投稿者'
          nullable: true

    RecipeInput:
      type: object
      description: レシピ作成・更新用の入力データ
      required:
        - title
        - roastLevel
      properties:
        title:
          type: string
          description: レシピのタイトル
          minLength: 1
          maxLength: 100
          example: 'エチオピア イルガチェフェ V60'
        summary:
          type: string
          description: レシピの概要説明
          minLength: 1
          maxLength: 500
          example: 'フルーティーで華やかな酸味が特徴的なエチオピア産コーヒー'
          nullable: true
        equipment:
          type: array
          description: 使用する器具のリスト
          items:
            type: string
            minLength: 1
            maxLength: 50
          minItems: 1
          maxItems: 10
          example: ['V60', 'コーヒーミル', 'ペーパーフィルター']
          nullable: true
        roastLevel:
          $ref: '#/components/schemas/RoastLevel'
        remarks:
          type: string
          description: 備考・コメント
          maxLength: 1000
          example: '特に蒸らし時間に注意。豆の膨らみを見て判断する。'
          nullable: true

    RecipeListResponse:
      type: object
      description: レシピリストのレスポンス
      required:
        - recipes
        - pagination
      properties:
        recipes:
          type: array
          description: レシピのリスト
          items:
            $ref: '#/components/schemas/Recipe'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      description: ページネーション情報
      required:
        - currentPage
        - totalPages
        - totalItems
        - itemsPerPage
      properties:
        currentPage:
          type: integer
          description: 現在のページ番号
          minimum: 1
          example: 1
        totalPages:
          type: integer
          description: 総ページ数
          minimum: 0
          example: 5
        totalItems:
          type: integer
          description: 総アイテム数
          minimum: 0
          example: 98
        itemsPerPage:
          type: integer
          description: 1ページあたりのアイテム数
          minimum: 1
          example: 20

    EquipmentResponse:
      type: object
      description: 器具データのレスポンス
      required:
        - equipment
      properties:
        equipment:
          $ref: '#/components/schemas/EquipmentByType'

    EquipmentByType:
      type: object
      description: タイプ別器具データ
      required:
        - grinder
        - dripper
        - filter
      properties:
        grinder:
          type: array
          description: コーヒーミルのリスト
          items:
            $ref: '#/components/schemas/EquipmentItem'
        dripper:
          type: array
          description: ドリッパーのリスト
          items:
            $ref: '#/components/schemas/EquipmentItem'
        filter:
          type: array
          description: ペーパーフィルターのリスト
          items:
            $ref: '#/components/schemas/EquipmentItem'

    EquipmentItem:
      type: object
      description: 器具アイテム
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          description: 器具の一意識別子
          example: '1'
        name:
          type: string
          description: 器具名
          minLength: 1
          maxLength: 50
          example: 'V60'
        brand:
          type: string
          nullable: true
          description: ブランド名
          minLength: 1
          maxLength: 50
          example: 'HARIO'
        type:
          type: string
          description: 器具のタイプ
          minLength: 1
          maxLength: 50
          example: 'ドリッパー'

    ErrorResponse:
      type: object
      description: 統一エラーレスポンス
      required:
        - error
        - message
        - request_id
        - timestamp
      properties:
        error:
          type: string
          description: エラータイプ（機械読み取り用）
          example: 'VALIDATION_ERROR'
        message:
          type: string
          description: エラーメッセージ（人間読み取り用）
          example: '入力データにエラーがあります'
        request_id:
          type: string
          description: リクエスト追跡ID
          example: 'req_abc123xyz789'
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻（ISO 8601形式）
        details:
          type: object
          nullable: true
          description: エラーの詳細情報

    RoastLevel:
      type: string
      description: コーヒー豆の焙煎度
      enum:
        - LIGHT
        - LIGHT_MEDIUM
        - MEDIUM
        - MEDIUM_DARK
        - DARK
        - FRENCH
      example: LIGHT_MEDIUM

    RecipeDetail:
      type: object
      description: レシピ詳細情報（バリスタ、手順、器具、タグを含む）
      required:
        - id
        - title
        - roastLevel
        - viewCount
        - isPublished
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: レシピの一意識別子
          example: '1'
        title:
          type: string
          description: レシピのタイトル
          minLength: 1
          maxLength: 100
          example: 'エチオピア イルガチェフェ V60'
        summary:
          type: string
          description: レシピの概要説明
          minLength: 1
          maxLength: 500
          example: 'フルーティーで華やかな酸味が特徴的なエチオピア産コーヒー'
          nullable: true
        remarks:
          type: string
          description: 備考・コメント
          maxLength: 1000
          example: '特に蒸らし時間に注意。豆の膨らみを見て判断する。'
          nullable: true
        roastLevel:
          $ref: '#/components/schemas/RoastLevel'
        viewCount:
          type: integer
          description: ビュー数
          minimum: 0
          example: 150
        isPublished:
          type: boolean
          description: 公開状態
          example: true
        publishedAt:
          type: string
          format: date-time
          description: 公開日時（ISO 8601形式）
          example: '2025-06-20T10:30:00Z'
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 作成日時（ISO 8601形式）
          example: '2025-06-20T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時（ISO 8601形式）
          example: '2025-06-22T10:30:00Z'
        barista:
          $ref: '#/components/schemas/Barista'
          nullable: true
          description: レシピを考案したバリスタ情報
        steps:
          type: array
          description: レシピの手順リスト
          items:
            $ref: '#/components/schemas/RecipeStep'
          example: []
        equipment:
          type: array
          description: 使用する器具の詳細リスト
          items:
            $ref: '#/components/schemas/DetailedEquipment'
          example: []
        tags:
          type: array
          description: レシピに関連するタグ
          items:
            $ref: '#/components/schemas/RecipeTag'
          example: []

    Barista:
      type: object
      description: バリスタ情報
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: バリスタの一意識別子
          example: '1'
        name:
          type: string
          description: バリスタ名
          minLength: 1
          maxLength: 100
          example: '佐藤花子'
        affiliation:
          type: string
          description: 所属店舗・組織
          maxLength: 200
          example: 'Specialty Coffee Shop ARIA'
          nullable: true
        socialLinks:
          type: array
          description: SNSリンク
          items:
            $ref: '#/components/schemas/SocialLink'
          example: []

    SocialLink:
      type: object
      description: SNSリンク情報
      required:
        - id
        - platform
        - url
      properties:
        id:
          type: string
          description: SNSリンクの一意識別子
          example: '1'
        platform:
          type: string
          description: SNSプラットフォーム名
          minLength: 1
          maxLength: 50
          example: 'Instagram'
        url:
          type: string
          format: uri
          description: SNSのURL
          example: 'https://instagram.com/hanako_barista'

    RecipeStep:
      type: object
      description: レシピの手順
      required:
        - id
        - stepOrder
        - description
      properties:
        id:
          type: string
          description: 手順の一意識別子
          example: '1'
        stepOrder:
          type: integer
          description: 手順の順序（1から開始）
          minimum: 1
          example: 1
        timeSeconds:
          type: integer
          description: 開始からの累積時間（秒）
          minimum: 0
          maximum: 3600
          example: 30
          nullable: true
        description:
          type: string
          description: 手順の詳細説明
          minLength: 1
          maxLength: 500
          example: 'お湯を沸かし、92℃まで冷ます。ドリッパーとサーバーを温めておく。'

    DetailedEquipment:
      type: object
      description: 詳細な器具情報
      required:
        - id
        - name
        - equipmentType
      properties:
        id:
          type: string
          description: 器具の一意識別子
          example: '1'
        name:
          type: string
          description: 器具名
          minLength: 1
          maxLength: 100
          example: 'V60ドリッパー'
        brand:
          type: string
          description: ブランド名
          maxLength: 100
          example: 'HARIO'
          nullable: true
        description:
          type: string
          description: 器具の説明
          maxLength: 500
          example: '円錐形で一つ穴のドリッパー。抽出のコントロールがしやすく、クリアな味わいが特徴。'
          nullable: true
        affiliateLink:
          type: string
          format: uri
          description: アフィリエイトリンク
          example: 'https://example.com/affiliate/hario-v60'
          nullable: true
        equipmentType:
          $ref: '#/components/schemas/EquipmentType'

    EquipmentType:
      type: object
      description: 器具タイプ情報
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: 器具タイプの一意識別子
          example: '1'
        name:
          type: string
          description: 器具タイプ名
          minLength: 1
          maxLength: 50
          example: 'ドリッパー'
        description:
          type: string
          description: 器具タイプの説明
          maxLength: 200
          example: 'コーヒーを抽出するための器具'
          nullable: true

    RecipeTag:
      type: object
      description: レシピタグ情報
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          description: タグの一意識別子
          example: '1'
        name:
          type: string
          description: タグ名
          minLength: 1
          maxLength: 50
          example: 'エチオピア'
        slug:
          type: string
          description: URL用のスラッグ
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 50
          example: 'ethiopia'

tags:
  - name: Recipes
    description: コーヒーレシピ関連のAPI
  - name: Equipment
    description: コーヒー器具関連のAPI